[DEBUG] edit data (ProductController)

+++ app/Http/Controllers/Backend/ProductController.php
@@ -209,24 +209,33 @@ class ProductController extends Controller
     //// Multi Image Update ////
     public function multiImageUpdate(Request $request){
 		$imgs = $request->multi_img;
-
-		foreach ($imgs as $id => $img) {
-            $imgDel = MultiImg::findOrFail($id);
-            unlink($imgDel->photo_name);
-
-            $make_name = hexdec(uniqid()).'.'.$img->getClientOriginalExtension();
-            Image::make($img)->resize(917,1000)->save('upload/products/multi-image/'.$make_name);
-            $uploadPath = 'upload/products/multi-image/'.$make_name;        
-            MultiImg::where('id',$id)->update([
-                'photo_name' => $uploadPath,
-                'updated_at' => Carbon::now(),
-            ]);
-        } // end foreach
-
-        $notification = array(
-			'message' => 'Product Image Updated Successfully',
-			'alert-type' => 'info'
-		);
+        if($imgs != NULL)
+        {
+            foreach ($imgs as $id => $img) {
+                $imgDel = MultiImg::findOrFail($id);
+                unlink($imgDel->photo_name);
+    
+                $make_name = hexdec(uniqid()).'.'.$img->getClientOriginalExtension();
+                Image::make($img)->resize(917,1000)->save('upload/products/multi-image/'.$make_name);
+                $uploadPath = 'upload/products/multi-image/'.$make_name;        
+                MultiImg::where('id',$id)->update([
+                    'photo_name' => $uploadPath,
+                    'updated_at' => Carbon::now(),
+                ]);
+            } // end foreach
+    
+            $notification = array(
+                'message' => 'Product Image Updated Successfully',
+                'alert-type' => 'info'
+            );
+        }
+        else
+        {
+            $notification = array(
+                'message' => 'No Image Insert',
+                'alert-type' => 'warning'
+            );
+        }
 		return redirect()->back()->with($notification);
 	} // end mehtod
